1.fix demonstring image error2.logout user3.add 404 page

+++ public/js/UserScript.VUE.js
@@ -1,9 +1,18 @@
 $(document).ready(function () {
 
-	const fetch_Users_Info = () => {
+
+	const sendToken = function (request) {
+		request.setRequestHeader(
+			'Authorization',
+			localStorage.getItem('x-auth') || sessionStorage.getItem("x-auth")
+		)
+	}
+
+
+	let fetch_Users_Info = () => {
 		$.get("/userList/fetch", function ({ data, count }) {
 
-			var table = new Vue({
+			let table = new Vue({
 				el: "#table",
 				data: {
 					people: data.reverse(),
@@ -12,22 +21,29 @@ $(document).ready(function () {
 				},
 				methods: {
 					//get userImageSource
-					getImageSource:function(user){
-						var address ="image/user/"+user.filename
+					getImageSource: function (user) {
+						let address = "image/user/" + user.filename
 						console.log(address);
 						return address
 					},
 					//remove Users
 					RemoveUser: function (user) {
+						if (user.userName === localStorage.getItem('userName') ||
+							user.userName === sessionStorage.getItem('userName')
+						) {
+							alert('please first logout and then remove yourself')
+							return
+						}
 						const result = confirm("you sure ?? you creazy???")
 						result
-						? 
+							?
 							$.ajax({
 								url: "/userList/removeUser",
 								method: "delete",
-								data:{_id : user._id}
+								data: { _id: user._id }
 							})
 								.done(function () {
+
 									alert(
 										`${user.userName}  has deleted now`
 									)
@@ -36,7 +52,7 @@ $(document).ready(function () {
 								.fail(function (err) {
 									alert(err.msg)
 								})
-							:()=>{
+							: () => {
 								return
 							}
 					},
@@ -61,7 +77,7 @@ $(document).ready(function () {
 
 	fetch_Users_Info()
 
-	var form = new Vue({
+	let form = new Vue({
 		el: ".form",
 		data: {
 			userName: null,
@@ -94,13 +110,13 @@ $(document).ready(function () {
 					})
 			},
 			AddNewUser: function () {
-				
-				var formdata = new FormData(document.getElementById('form-data'))
 
-				formdata.append('userName' , this.userName)
-				formdata.append('phone' , this.phone)
-				formdata.append('email' , this.email)
-				formdata.append('password' , this.password)
+				let formdata = new FormData(document.getElementById('form-data'))
+
+				formdata.append('userName', this.userName)
+				formdata.append('phone', this.phone)
+				formdata.append('email', this.email)
+				formdata.append('password', this.password)
 
 
 				// SIGNUP NEW USER
@@ -108,16 +124,16 @@ $(document).ready(function () {
 					method: "post",
 					url: "/userList/signUp",
 					processData: false,
-					contentType: false,					
+					contentType: false,
 					data: formdata,
-					success:function(res){
-						Success=true
+					success: function (res) {
+						Success = true
 						login.saveUserToken(res)
 						location.assign(`/userList/?msg=${res.msg}`)
 					},
-					error:function(err){
-						Success=false
-						console.log(err)
+					error: function (err) {
+						Success = false
+						console.error(err)
 					}
 				})
 			}
@@ -144,14 +160,13 @@ $(document).ready(function () {
 					method: "post",
 					url: "/userList/login",
 					data: info,
-					success:function(data){
-						Success=true;
+					success: function (data) {
+						Success = true;
 						// SET TOKEN IN LOCAL STORAGE OR SESSION STORAGE
 						thiss.saveUserToken(data, thiss.permanent)
-						window.location.assign(`/userList/?msg=${data.msg}`)
 					},
-					error:function(err){
-						Success=false;
+					error: function (err) {
+						Success = false;
 						console.error(err)
 
 					}
@@ -162,18 +177,50 @@ $(document).ready(function () {
 				sessionStorage.clear()
 				if (permanent) {
 
-					localStorage.setItem("x-auth",data.token)
-					localStorage.setItem("imageFilename",data.imageFilename)
-					localStorage.setItem("userName",data.userName)
-					localStorage.setItem("cash",data.cash)
-				
-				}else{
-					sessionStorage.setItem("x-auth",data.token)
-					sessionStorage.setItem("imageFilename",data.imageFilename)
-					sessionStorage.setItem("userName",data.userName)
-					sessionStorage.setItem("cash",data.cash)
+					localStorage.setItem("x-auth", data.token)
+					localStorage.setItem("imageFilename", data.imageFilename)
+					localStorage.setItem("userName", data.userName)
+					localStorage.setItem("cash", data.cash)
+
+				} else {
+					sessionStorage.setItem("x-auth", data.token)
+					sessionStorage.setItem("imageFilename", data.imageFilename)
+					sessionStorage.setItem("userName", data.userName)
+					sessionStorage.setItem("cash", data.cash)
 				}
 			}
 		}
 	})
+
+
+	let token = localStorage.getItem('x-auth') || sessionStorage.getItem("x-auth")
+
+	if (token === null) {
+		$("#logout_button").hide()
+	} else {
+		$("#logout_button").on("click", function () {
+			let result = confirm("Are you sure?? do you want logout now")
+			result
+			?
+				$.ajax({
+					beforeSend: sendToken,
+					method: 'post',
+					url: "/userList/logout",
+					success: function (result) {
+						localStorage.clear()
+						sessionStorage.clear()
+						console.log(result);
+						window.location.reload()
+						$("#logout_button").hide()
+					},
+					error: function (error) {
+						console.log(error)
+					}
+				})
+			: 
+				(function(){
+					return
+				})()
+		})
+	}
 })
